fix: double-operation and add status icons
