+++ src/components/DocumentCard/core/TimeAgo.tsx
@@ -1,0 +1,11 @@
+import {useTimeAgo} from 'sanity'
+
+export interface TimeAgoProps {
+  time: string | Date
+}
+
+export function TimeAgo({time}: TimeAgoProps) {
+  const timeAgo = useTimeAgo(time)
+
+  return <span title={timeAgo}>{timeAgo} ago</span>
+}
