+++ src/components/WorkflowTool.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import {Flex, Card, Grid, Spinner, Label, useToast, Container, useTheme, Button} from '@sanity/ui'
+import {Box, Flex, Card, Grid, Spinner, Label, Container, useTheme, Button} from '@sanity/ui'
 import {Feedback, useProjectUsers} from 'sanity-plugin-utils'
-import {Tool, useClient, UserAvatar, useSchema} from 'sanity'
+import {Tool, UserAvatar, useSchema} from 'sanity'
 import {ResetIcon} from '@sanity/icons'
 import {DragDropContext, Droppable, Draggable, DropResult} from 'react-beautiful-dnd'
 
@@ -9,12 +9,14 @@ import {SanityDocumentWithMetadata, State, WorkflowConfig} from '../types'
 import {DocumentCard} from './DocumentCard'
 import {useWorkflowDocuments} from '../hooks/useWorkflowDocuments'
 import {API_VERSION, ORDER_MAX, ORDER_MIN} from '../constants'
-import FloatingCard from './FloatingCard'
+
+import Validators from './Validators'
 
 function filterItemsByStateAndUserAndSort(
   items: SanityDocumentWithMetadata[],
   stateId: string,
-  selectedUsers: string[]
+  selectedUsers: string[],
+  selectedSchemaTypes: string[]
 ) {
   return items
     .filter((item) => item?._metadata?.state === stateId)
@@ -22,6 +24,11 @@ function filterItemsByStateAndUserAndSort(
       selectedUsers.length && item._metadata?.assignees.length
         ? item._metadata?.assignees.some((assignee) => selectedUsers.includes(assignee))
         : !selectedUsers.length
+    )
+    .filter((item) =>
+      selectedSchemaTypes.length
+        ? selectedSchemaTypes.includes(item._type)
+        : Boolean(selectedSchemaTypes.length)
     )
     .sort((a, b) => {
       const aOrder = a?._metadata?.order ?? 0
@@ -38,10 +45,7 @@ type WorkflowToolProps = {
 export default function WorkflowTool(props: WorkflowToolProps) {
   const {schemaTypes = [], states = []} = props?.tool?.options ?? {}
 
-  const client = useClient({apiVersion: API_VERSION})
-  const toast = useToast()
   const schema = useSchema()
-
   const isDarkMode = useTheme().sanity.color.dark
   const defaultCardTone = isDarkMode ? 'default' : 'transparent'
 
@@ -53,141 +57,6 @@ export default function WorkflowTool(props: WorkflowToolProps) {
 
   // Operations to perform on cards
   const {move} = operations
-
-  // A lot of error-checking
-  const documentsWithoutMetadataIds = data?.length
-    ? data.filter((doc) => !doc._metadata).map((d) => d._id.replace(`drafts.`, ``))
-    : []
-  const documentsWithoutValidMetadataIds = data?.length
-    ? data.reduce((acc, cur) => {
-        const {documentId, state} = cur._metadata ?? {}
-        const stateExists = states.find((s) => s.id === state)
-
-        return !stateExists && documentId ? [...acc, documentId] : acc
-      }, [] as string[])
-    : []
-  const documentsWithoutValidUsersIds = data?.length
-    ? data.reduce((acc, cur) => {
-        const {documentId, assignees} = cur._metadata ?? {}
-        const assigneesExist = assignees?.every((a) => userList.find((u) => u.id === a))
-
-        return !assigneesExist && documentId ? [...acc, documentId] : acc
-      }, [] as string[])
-    : []
-  const documentsWithoutOrderIds = data?.length
-    ? data.reduce((acc, cur) => {
-        const {documentId, order} = cur._metadata ?? {}
-
-        return !order && documentId ? [...acc, documentId] : acc
-      }, [] as string[])
-    : []
-
-  // Creates metadata documents for those that do not have them
-  const importDocuments = React.useCallback(
-    async (ids: string[]) => {
-      toast.push({
-        title: 'Importing documents',
-        status: 'info',
-      })
-
-      const tx = ids.reduce((item, documentId) => {
-        return item.createOrReplace({
-          _id: `workflow-metadata.${documentId}`,
-          _type: 'workflow.metadata',
-          state: states[0].id,
-          documentId,
-        })
-      }, client.transaction())
-
-      await tx.commit()
-
-      toast.push({
-        title: `Imported ${ids.length === 1 ? `1 Document` : `${ids.length} Documents`}`,
-        status: 'success',
-      })
-    },
-    [client, states, toast]
-  )
-
-  // Updates metadata documents to a valid, existing state
-  const correctDocuments = React.useCallback(
-    async (ids: string[]) => {
-      toast.push({
-        title: 'Correcting...',
-        status: 'info',
-      })
-
-      const tx = ids.reduce((item, documentId) => {
-        return item.patch(`workflow-metadata.${documentId}`, {
-          set: {state: states[0].id},
-        })
-      }, client.transaction())
-
-      await tx.commit()
-
-      toast.push({
-        title: `Corrected ${ids.length === 1 ? `1 Document` : `${ids.length} Documents`}`,
-        status: 'success',
-      })
-    },
-    [client, states, toast]
-  )
-
-  // Remove users that are no longer in the project from documents
-  const removeUsersFromDocuments = React.useCallback(
-    async (ids: string[]) => {
-      toast.push({
-        title: 'Removing users...',
-        status: 'info',
-      })
-
-      const tx = ids.reduce((item, documentId) => {
-        const {assignees} = data.find((d) => d._id === documentId)?._metadata ?? {}
-        const validAssignees = assignees?.length
-          ? // eslint-disable-next-line max-nested-callbacks
-            assignees.filter((a) => userList.find((u) => u.id === a)?.id)
-          : []
-
-        return item.patch(`workflow-metadata.${documentId}`, {
-          set: {assignees: validAssignees},
-        })
-      }, client.transaction())
-
-      await tx.commit()
-
-      toast.push({
-        title: `Corrected ${ids.length === 1 ? `1 Document` : `${ids.length} Documents`}`,
-        status: 'success',
-      })
-    },
-    [client, data, toast]
-  )
-
-  // Add order value to metadata documents
-  const addOrderToDocuments = React.useCallback(
-    async (ids: string[]) => {
-      toast.push({
-        title: 'Adding ordering...',
-        status: 'info',
-      })
-
-      // TODO: This doesn't consider the order of other documents
-      const startingValue = 10000
-      const tx = ids.reduce((item, documentId, itemIndex) => {
-        return item.patch(`workflow-metadata.${documentId}`, {
-          set: {order: startingValue + itemIndex * 1000},
-        })
-      }, client.transaction())
-
-      await tx.commit()
-
-      toast.push({
-        title: `Added order to ${ids.length === 1 ? `1 Document` : `${ids.length} Documents`}`,
-        status: 'success',
-      })
-    },
-    [client, toast]
-  )
 
   const handleDragEnd = React.useCallback(
     (result: DropResult) => {
@@ -204,7 +73,7 @@ export default function WorkflowTool(props: WorkflowToolProps) {
 
       // Find all items in current state
       const destinationStateItems = [
-        ...filterItemsByStateAndUserAndSort(data, destination.droppableId, []),
+        ...filterItemsByStateAndUserAndSort(data, destination.droppableId, [], []),
       ]
       let newOrder = ORDER_MIN
 
@@ -243,10 +112,18 @@ export default function WorkflowTool(props: WorkflowToolProps) {
 
     return userList.filter((user) => uniqueUserIds.includes(user.id))
   }, [data, userList])
+
   const [selectedUsers, setSelectedUsers] = React.useState<any[]>(uniqueAssignedUsers)
   const toggleSelectedUser = React.useCallback((userId: string) => {
     setSelectedUsers((prev) =>
       prev.includes(userId) ? prev.filter((u) => u !== userId) : [...prev, userId]
+    )
+  }, [])
+
+  const [selectedSchemaTypes, setSelectedSchemaTypes] = React.useState<string[]>(schemaTypes)
+  const toggleSelectedSchemaType = React.useCallback((schemaType: string) => {
+    setSelectedSchemaTypes((prev) =>
+      prev.includes(schemaType) ? prev.filter((u) => u !== schemaType) : [...prev, schemaType]
     )
   }, [])
 
@@ -272,53 +149,8 @@ export default function WorkflowTool(props: WorkflowToolProps) {
 
   return (
     <>
-      <FloatingCard>
-        {documentsWithoutMetadataIds.length > 0 ? (
-          <Button
-            tone="caution"
-            onClick={() => importDocuments(documentsWithoutMetadataIds)}
-            text={
-              documentsWithoutMetadataIds.length === 1
-                ? `Import 1 Missing Document into Workflow`
-                : `Import ${documentsWithoutMetadataIds.length} Missing Documents into Workflow`
-            }
-          />
-        ) : null}
-        {documentsWithoutValidMetadataIds.length > 0 ? (
-          <Button
-            tone="caution"
-            onClick={() => correctDocuments(documentsWithoutValidMetadataIds)}
-            text={
-              documentsWithoutValidMetadataIds.length === 1
-                ? `Correct 1 Document State`
-                : `Correct ${documentsWithoutValidMetadataIds.length} Document States`
-            }
-          />
-        ) : null}
-        {documentsWithoutValidUsersIds.length > 0 ? (
-          <Button
-            tone="caution"
-            onClick={() => removeUsersFromDocuments(documentsWithoutValidUsersIds)}
-            text={
-              documentsWithoutValidUsersIds.length === 1
-                ? `Remove Invalid Users from 1 Document`
-                : `Remove Invalid Users from ${documentsWithoutValidUsersIds.length} Documents`
-            }
-          />
-        ) : null}
-        {documentsWithoutOrderIds.length > 0 ? (
-          <Button
-            tone="caution"
-            onClick={() => addOrderToDocuments(documentsWithoutOrderIds)}
-            text={
-              documentsWithoutOrderIds.length === 1
-                ? `Set Order for 1 Document`
-                : `Set Order for ${documentsWithoutOrderIds.length} Documents`
-            }
-          />
-        ) : null}
-      </FloatingCard>
-      <Card tone="primary" padding={2} borderBottom>
+      <Validators data={data} userList={userList} states={states} />
+      <Card tone="primary" padding={2} borderBottom style={{overflowX: 'hidden'}}>
         <Flex justify="space-between">
           {uniqueAssignedUsers.length > 0 ? (
             <Flex align="center" gap={1}>
@@ -334,19 +166,19 @@ export default function WorkflowTool(props: WorkflowToolProps) {
               ))}
 
               {selectedUsers.length > 0 ? (
-                <Card borderLeft marginLeft={2} paddingLeft={2} tone="inherit">
+                <Card borderLeft marginLeft={2} paddingLeft={3} tone="inherit">
                   <Button
                     text="Clear"
                     onClick={() => setSelectedUsers([])}
-                    // fontSize={1}
-                    // padding={2}
                     mode="ghost"
                     icon={ResetIcon}
                   />
                 </Card>
               ) : null}
             </Flex>
-          ) : null}
+          ) : (
+            <Box flex={1} />
+          )}
           {schemaTypes.length > 0 ? (
             <Flex align="center" gap={1}>
               {schemaTypes.map((type) => (
@@ -354,6 +186,9 @@ export default function WorkflowTool(props: WorkflowToolProps) {
                   key={type}
                   text={schema.get(type)?.title ?? type}
                   icon={schema.get(type)?.icon ?? undefined}
+                  mode={selectedSchemaTypes.includes(type) ? `default` : `ghost`}
+                  fontSize={2}
+                  onClick={() => toggleSelectedSchemaType(type)}
                 />
               ))}
             </Flex>
@@ -381,31 +216,34 @@ export default function WorkflowTool(props: WorkflowToolProps) {
                     ) : null}
 
                     {data.length > 0 &&
-                      filterItemsByStateAndUserAndSort(data, state.id, selectedUsers).map(
-                        (item, itemIndex) => (
-                          // The metadata's documentId is always the published one
-                          <Draggable
-                            key={item?._metadata?.documentId as string}
-                            draggableId={item?._metadata?.documentId as string}
-                            index={itemIndex}
-                          >
-                            {(draggableProvided, draggableSnapshot) => (
-                              <div
-                                ref={draggableProvided.innerRef}
-                                {...draggableProvided.draggableProps}
-                                {...draggableProvided.dragHandleProps}
-                              >
-                                <DocumentCard
-                                  isDragging={draggableSnapshot.isDragging}
-                                  item={item}
-                                  userList={userList}
-                                  states={states}
-                                />
-                              </div>
-                            )}
-                          </Draggable>
-                        )
-                      )}
+                      filterItemsByStateAndUserAndSort(
+                        data,
+                        state.id,
+                        selectedUsers,
+                        selectedSchemaTypes
+                      ).map((item, itemIndex) => (
+                        // The metadata's documentId is always the published one
+                        <Draggable
+                          key={item?._metadata?.documentId as string}
+                          draggableId={item?._metadata?.documentId as string}
+                          index={itemIndex}
+                        >
+                          {(draggableProvided, draggableSnapshot) => (
+                            <div
+                              ref={draggableProvided.innerRef}
+                              {...draggableProvided.draggableProps}
+                              {...draggableProvided.dragHandleProps}
+                            >
+                              <DocumentCard
+                                isDragging={draggableSnapshot.isDragging}
+                                item={item}
+                                userList={userList}
+                                states={states}
+                              />
+                            </div>
+                          )}
+                        </Draggable>
+                      ))}
                     {provided.placeholder}
                   </Card>
                 )}
